+++ TheExpanse/Vector.h
@@ -4,10 +4,10 @@ class Vector
 private:
 
 public:
-	int x;
-	int y;
+	double x;
+	double y;
 	Vector();
-	Vector(float x, float y);
+	Vector(double x, double y);
 	Vector operator+(const Vector& operand);
 	Vector operator-(const Vector& operand);
 	Vector operator*(const Vector& operand);
